Fix test-build.sh for clang15
With clang15 the following warning is emitted when linking:

  clang-15: error: '-x c++' after last input file has no effect [-Werror,-Wunused-command-line-argument]

The problem is that automake adds the CFLAGS variable to both the
COMPILE and LINK variables. However -xc++ should be only used in the
COMPILE variable and clang15 now warns about this. The solution is to
set it in the CPPFLAGS variable instead which is only added to COMPILE.